update cell languages after each execution
