+++ src/dotnet-interactive-vscode/package.json
@@ -274,7 +274,7 @@
     "mocha-multi-reporters": "1.1.7",
     "mocha-trx-reporter": "3.2.4",
     "tmp": "0.2.1",
-    "typescript": "3.8.3",
+    "typescript": "4.1.3",
     "vsce": "1.75.0",
     "vscode-oniguruma": "1.3.1",
     "vscode-test": "1.3.0",
